requirements.txt
aio-pika==9.0.7
aiohttp==3.8.4
aiormq==6.7.6
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
astor==0.8.1
async-timeout==4.0.2
attrs==23.1.0
bandit==1.7.5
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
darglint==1.8.1
Django==3.2
django-debug-toolbar==3.4.0
django-split-settings==1.2.0
djangorestframework==3.14.0
docutils==0.20
drf-spectacular==0.26.2
eradicate==2.2.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.12.6
flake8-commas==2.1.0
flake8-comprehensions==3.12.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.4.0
flake8-html==0.4.3
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.6.1
mdurl==0.1.2
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
pamqp==3.2.1
pbr==5.11.1
pep8-naming==0.13.2
pika==1.3.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9
pycodestyle==2.8.0
pydantic==1.10.7
pydocstyle==6.3.0
pyflakes==2.4.0
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
redis==4.5.5
requests==2.30.0
restructuredtext-lint==1.4.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
spectree==1.1.2
SQLAlchemy==2.0.14
sqlparse==0.4.3
stevedore==5.0.0
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
validate-email==1.3
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
wemake-python-styleguide==0.17.0
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
gunicorn==20.1.0